creative-ads-platform/dashboard/backend/requirements.txt
# FastAPI Backend for Creative Ads Dashboard
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-bigquery==3.14.1
google-cloud-monitoring==2.19.0
google-cloud-logging==3.9.0
google-auth==2.27.0

# Vertex AI
google-cloud-aiplatform==1.39.0
vertexai==1.39.0

# Async Support
aiohttp==3.9.1
aiocache==0.12.2

# Database
aiosqlite==0.19.0

# Utils
python-dateutil==2.8.2
structlog==24.1.0
prometheus-client==0.19.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

